random: make crypto-generate-random-string more idiomatic
It's about twice as fast now.  The next step to make it faster would
be to cache larger chunks of crypto random bytes into memory.  A
single call to `crypto-random-bytes' on my machine takes 130
microseconds, which means we can now only generate about 10 of these
per millisecond, which isn't great.  By amortizing the cost of going
out to the filesystem over time, we can probably get this to run a
couple orders of magnitude faster.